MINOR: move the test cases which don't need brokers from TopicCommandWithAdminClientTest to TopicCommandTest
