+++ core/src/test/scala/unit/kafka/admin/TopicCommandWithAdminClientTest.scala
@@ -19,10 +19,9 @@ package kafka.admin
 import java.util.{Collections, Optional, Properties}
 
 import kafka.admin.TopicCommand.{AdminClientTopicService, TopicCommandOptions}
-import kafka.common.AdminCommandFailedException
 import kafka.integration.KafkaServerTestHarness
 import kafka.server.{ConfigType, KafkaConfig}
-import kafka.utils.{Exit, Logging, TestUtils}
+import kafka.utils.{Logging, TestUtils}
 import kafka.zk.{ConfigEntityChangeNotificationZNode, DeleteTopicsTopicZNode}
 import org.apache.kafka.clients.CommonClientConfigs
 import org.apache.kafka.clients.admin._
@@ -79,33 +78,14 @@ class TopicCommandWithAdminClientTest extends KafkaServerTestHarness with Loggin
   private var testTopicName: String = _
 
   private val _testName = new TestName
-  @Rule def testName = _testName
+  @Rule def testName: TestName = _testName
 
-  def assertExitCode(expected: Int, method: () => Unit): Unit = {
-    def mockExitProcedure(exitCode: Int, exitMessage: Option[String]): Nothing = {
-      assertEquals(expected, exitCode)
-      throw new RuntimeException
-    }
-    Exit.setExitProcedure(mockExitProcedure)
-    try {
-      intercept[RuntimeException] {
-        method()
-      }
-    } finally {
-      Exit.resetExitProcedure()
-    }
-  }
-
-  def assertCheckArgsExitCode(expected: Int, options: TopicCommandOptions): Unit = {
-    assertExitCode(expected, options.checkArgs _)
-  }
-
-  def createAndWaitTopic(opts: TopicCommandOptions): Unit = {
+  private[this] def createAndWaitTopic(opts: TopicCommandOptions): Unit = {
     topicService.createTopic(opts)
     waitForTopicCreated(opts.topic.get)
   }
 
-  def waitForTopicCreated(topicName: String, timeout: Int = 10000): Unit = {
+  private[this] def waitForTopicCreated(topicName: String, timeout: Int = 10000): Unit = {
     TestUtils.waitUntilMetadataIsPropagated(servers, topicName, partition = 0, timeout)
   }
 
@@ -124,42 +104,6 @@ class TopicCommandWithAdminClientTest extends KafkaServerTestHarness with Loggin
     // adminClient is closed by topicService
     if (topicService != null)
       topicService.close()
-  }
-
-  @Test
-  def testParseAssignment(): Unit = {
-    val actualAssignment = TopicCommand.parseReplicaAssignment("5:4,3:2,1:0")
-    val expectedAssignment = Map(0 -> List(5, 4), 1 -> List(3, 2), 2 -> List(1, 0))
-    assertEquals(expectedAssignment, actualAssignment)
-  }
-
-  @Test
-  def testParseAssignmentDuplicateEntries(): Unit = {
-    intercept[AdminCommandFailedException] {
-      TopicCommand.parseReplicaAssignment("5:5")
-    }
-  }
-
-  @Test
-  def testParseAssignmentPartitionsOfDifferentSize(): Unit = {
-    intercept[AdminOperationException] {
-      TopicCommand.parseReplicaAssignment("5:4:3,2:1")
-    }
-  }
-
-  @Test
-  def testConfigOptWithBootstrapServers(): Unit = {
-    assertCheckArgsExitCode(1,
-      new TopicCommandOptions(Array("--bootstrap-server", brokerList ,"--alter", "--topic", testTopicName, "--partitions", "3", "--config", "cleanup.policy=compact")))
-    assertCheckArgsExitCode(1,
-      new TopicCommandOptions(Array("--bootstrap-server", brokerList ,"--alter", "--topic", testTopicName, "--partitions", "3", "--delete-config", "cleanup.policy")))
-    val opts =
-      new TopicCommandOptions(Array("--bootstrap-server", brokerList ,"--create", "--topic", testTopicName, "--partitions", "3", "--replication-factor", "3", "--config", "cleanup.policy=compact"))
-    opts.checkArgs()
-    assertTrue(opts.hasCreateOption)
-    assertEquals(brokerList, opts.bootstrapServer.get)
-    assertEquals("cleanup.policy=compact", opts.topicConfig.get.get(0))
-
   }
 
   @Test
@@ -283,28 +227,6 @@ class TopicCommandWithAdminClientTest extends KafkaServerTestHarness with Loggin
       topicService.createTopic(new TopicCommandOptions(
         Array("--partitions", "2", "--replication-factor", "-1", "--topic", testTopicName)))
     }
-  }
-
-  @Test
-  def testCreateWithAssignmentAndPartitionCount(): Unit = {
-    assertCheckArgsExitCode(1,
-      new TopicCommandOptions(
-        Array("--bootstrap-server", brokerList,
-          "--create",
-          "--replica-assignment", "3:0,5:1",
-          "--partitions", "2",
-          "--topic", "testTopic")))
-  }
-
-  @Test
-  def testCreateWithAssignmentAndReplicationFactor(): Unit = {
-    assertCheckArgsExitCode(1,
-      new TopicCommandOptions(
-        Array("--bootstrap-server", brokerList,
-          "--create",
-          "--replica-assignment", "3:0,5:1",
-          "--replication-factor", "2",
-          "--topic", "testTopic")))
   }
 
   @Test
@@ -434,12 +356,6 @@ class TopicCommandWithAdminClientTest extends KafkaServerTestHarness with Loggin
       topicService.alterTopic(new TopicCommandOptions(
         Array("--partitions", "-1", "--topic", testTopicName)))
     }
-  }
-
-  @Test
-  def testAlterWithUnspecifiedPartitionCount(): Unit = {
-    assertCheckArgsExitCode(1, new TopicCommandOptions(
-      Array("--bootstrap-server", brokerList ,"--alter", "--topic", testTopicName)))
   }
 
   @Test
@@ -614,9 +530,7 @@ class TopicCommandWithAdminClientTest extends KafkaServerTestHarness with Loggin
               val testPartitionMetadata = topicMetadatas.find(_.name.equals(testTopicName)).get.partitions.asScala.find(_.partitionIndex == partitionOnBroker0)
               testPartitionMetadata match {
                 case None => fail(s"Partition metadata is not found in metadata cache")
-                case Some(metadata) => {
-                  result && metadata.errorCode == Errors.LEADER_NOT_AVAILABLE.code
-                }
+                case Some(metadata) => result && metadata.errorCode == Errors.LEADER_NOT_AVAILABLE.code
               }
             }
           }
@@ -744,7 +658,7 @@ class TopicCommandWithAdminClientTest extends KafkaServerTestHarness with Loggin
         topicService.describeTopic(new TopicCommandOptions(Array("--at-min-isr-partitions"))))
       val rows = output.split("\n")
       assertTrue(rows(0).startsWith(s"\tTopic: $testTopicName"))
-      assertEquals(1, rows.length);
+      assertEquals(1, rows.length)
     } finally {
       restartDeadBrokers()
     }
@@ -790,7 +704,7 @@ class TopicCommandWithAdminClientTest extends KafkaServerTestHarness with Loggin
       val rows = output.split("\n")
       assertTrue(rows(0).startsWith(s"\tTopic: $underMinIsrTopic"))
       assertTrue(rows(1).startsWith(s"\tTopic: $offlineTopic"))
-      assertEquals(2, rows.length);
+      assertEquals(2, rows.length)
     } finally {
       restartDeadBrokers()
     }
